Item id signature to 16bit long & indentation typo

+++ src/items.h
@@ -207,7 +207,7 @@ class ItemType
 		int32_t defense = 0;
 		int32_t extraDefense = 0;
 		int32_t armor = 0;
-		int32_t rotateTo = 0;
+		uint16_t rotateTo = 0;
 		int32_t runeMagLevel = 0;
 		int32_t runeLevel = 0;
